Histalign/Etc: Add `AlignmentParameterAggregator` to save registration
This `AlignmentParameterAggregator` is a model that keeps track of both
the `VolumeSettings` and `HistologySettings` changes as well as the
scaling factors that were applied to the volume and histological image
to fit them on screen.
Finally, it keeps track of the pixels sizes of both images so that
downsampling can be compensated for. Since the two images are both using
(0, 0) as their center point, the resolution can be compensated for by
reconstructing adequate transformations when doing quantification based
on the registration parameters.