Histalign/Application: Add workspace functionality
The two classes added with this commit, `Workspace` and
`HistologySlice`, both work in tandem to provide an easy interface to
work with directories of related histology images.

The `Workspace` class acts as a manager of `HistologySlice`s, it keeps
track of them in order and handles the interface to load and iterate
over them.

The `HistologySlice` provides a wrapper around images on the file system
by handling reading from disk and the generation of thumbnails for the
UI. By default, thumbnails are not generated when the `HistologySlice`
object is created as that would require loading the image into memory.
However, if provided a working directory during thumbnail generation,
the thumbnail will automatically be cached.