Histalign/Application: Remove the need to provide a slice on startup
Now that we have a workspace and thumbnails that the user can interact
with, it is no longer necessary to give the application a slice to have
open when starting it.
Instead, the user can now double click on any of the thumbnails to have
that open in the alignment view.