Histalign/Application: Make GUI more resilient to starting empty
This adds some placeholders in the thumbnail dock widget so that it
still keeps its shape even when empty (i.e., on startup).
It also decouples it from keeping track of the workspace as it only
cares about it for painting thumbnails, and it is now the only
interaction it has with it rather than keeping it as an attribute.