Histalign/Application: Make `Workspace` a QObject and improve thumbnails
`Workspace` is now a QObject so that it can make use of signals and
slots, making it much easier to synchronise the thumbnail UI with the
thumbnails that have been generated/loaded.

There is still the possibility of a thumbnail being dropped while it is
being processed by the UI (e.g., when changing image directory while
generating thumbnails for a previous one), so the thumbnails are copied
before being emitted by the `Workspace` signal.