Histalign/Workspace: Distinguish between project and working directories
The project directory is set as the root of the current project. It will
hold the relevant information about the current state of the project or
the UI (i.e., what the user would expect to have loaded when resuming
registration after closing and reopening the software) inside the
`project.txt` file.
The current working directory is the directory for the current image
directory. It holds the thumbnail cache and stores the registration
results. The directory name is taken from the hash of the image
directory after it has been resolved. This should make it unique
regardless of whether symlinks were used to open the image directory.