Meta: Restructure codebase
Now that it is a bit clearer how things interact with each other, this
refactor organises the code base in a clearer manner.
The modules are now split between the front end and the back end and
the application itself is set up to handle different GUIs. This will
make it easier in the future to create new GUIs for different things
(e.g., a separate GUI and command for starting registration and
visualisation).