Histalign/Workspace: Clean up RegistrationMainWindow and Workspace
This completes the restructure of the codebase.
The systems should be modular enough to allow easier improvements in the
future while maintaining a good enough separation of responsibility.

I am still learning how to structure a GUI application code-wise so this
might change again in the future but it seems organised enough in this
state.