Histalign/ThumbnailScrollArea: Make vertical scrollbar always on
There's an annoying bug where the scroll area having a size that is just
at the point where without a vertical scrollbar, the thumbnails take
enough height to require one but with one, they don't need it. So the
resizeEvent is triggered in an infinite loop. It's ugly but having the
scrollbar always on means we avoid this.