Histalign/Quantification: Add prototype quantifier
The quantifier takes an output directory from the registration step and
carries out quantification off the structures provided.
It can also be told to use fast rescale and/or fast transform. The
"slow" approaches use `skimage`'s functions which are very accurage.
The `fast_rescale` flag makes the quantifier use `PIL`'s rescale which
might have some big changes on colour images but since we're using it
to rescale a mask, it seems equivalent to `skimage`'s and is therefore
the default.
The `fast_transform` flag makes the quantifier use `OpenCV`'s
`warpPerspective` instead of `skimage`'s `warp`. Once again, it loses
som accuracy (as it uses a smaller datatype) but is something like 20x
faster.

Currently, only average fluorescence is available but the quantifier
should be "plug and play" for other quantification methods.