Histalign/IO: Try to handle cases with missing volume files
Not all structures have a mask available from the Allen institute's
server.
This should be much of a problem in most cases and should only show up
in cases where quantification is attempted on such structures, as well
as when the user tries to show a QA contour of one of those structures.

This patch isn't perfect by any means but at least we plan for the
errors and notify the user of what went wrong.