Histalign/StructureHierarchyModel: Add parameter to select root node
Instead of keeping all of the parsed tree as a child of the
invisibleRootItem, the model now allows a `root_text` argument to be
passed to select which node to have as the root. This is particularly
helpful to select the one tree we're usually interested in ("Basic cell
groups and regions") while keeping the option to select tracts, etc if
they ever become relevant.

The tree is still parsed in its entirety and the nodes are only
re-parented after parsing. The code seems more readable this way and the
cost of parsing such a small tree is pretty negligible.