Histalign/Workspace: Refactor workspace objects
The various workspace-related objects have now been refactored into a
coherent naming scheme and their relationships have been reworked.

`Volume` is a new wrapper around a `vedo.Volume` that allows lazy
loading (and downloading) of CCF volumes. It is eager by default but
can be made lazy through an initialiser parameter.

`VolumeManager` has been renamed to `VolumeSlicer` as that is its
primary purpose. It now works with a `Volume` to lazily load its
resource.
This puts it in an easy position to make use of `VolumeLoaderThread`
(previously `AtlasHandler`) to load its volume asynchronously.

As an added bonus, I have dropped my habit of creating a file for each
class.