Histalign/CLI+Frontend: Add new centralised UI
This provides a simple starting point when starting the application for
the first time.
The user is given a `QTabWidget` with one tab per app functionality.
From there, they can choose to use any of them depending on their needs.

The backbone of this switching of windows is done through the
`ApplicationWidget` which is made the direct parent of any new window
when switching between them. This avoids fighting the garbage collector
to maintain unrelated windows.

Additionally, the user is also able to start any of the available UIs
through the CLI thanks to `click` grouping.