Histalign/Quantification: First draft for quantification GUI
The user loads a project like with other GUIs.
They are given a list of the directories that have been aligned as well
as a few parameters to control what the quantification will run on:
- the method: what measure to use,
- the structures: what structures to measure,
- whether to perform fast rescaling and transformation.

There are two other tabs: results and view. Results will be used to
display all the previously run quantifications while the view tab will
be used to display the actual data of the selected results.