Histalign: Fix offset calculations
I believe this should fix the offset calculations which were off by one.
Offsets are now allowed to have values in the range:
-(dimension // 2) to (dimension // 2 + int(dimension % 2) - 1).

For example, for an axis of length 10, an offset of zero represents the
0-value in a similar way to 0 representing the 0-value in an int8.
Therefore, the offset value should only be valid if it is in the bounds
presented above.
With an example of an array of length 10. O offset points to index 5:
| 0 | 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 |
 -5  -4  -3  -2  -1   0   1   2   3   4
With an example of an array of length 11. O offset points to index 5:
| 0 | 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 | 10 |
 -5  -4  -3  -2  -1   0   1   2   3   4    5