Histalign/Visualisation+IO: Add visualisation skeleton and IO helpers
The IO helpers help with converting numpy arrays of images from 8-bit
to colour, as well as masking a particular colour (e.g., mask off blue
to create yellow instead of gray).

The visualisation UI takes the form of something resembling an FMRI
viewer, showing all three main orientations at the same time with an
overlay of the aligned volume.
Currently, only the slices are presented as they are in the alignment
volume. In the future, this should be converted to a heatmap-like
volume (e.g., for average fluorescence).