Histalign/Meta: Remove a lot of boilerplate code and add to BasicMenuBar
There is now a `HistalignMainWindow` parent class for all the individual
main windows to inherit from. This implements the boilerplate code
related to creating a menu bar, showing the open project dialog, and
closing the window to return to the centralised one.
Note that inheriting from ABC is very difficult with PySide so abstract
methods are simply decorated but ABC is not inherited. This helps IDEs
present the problem to the user but there is no runtime check on
unimplemented abstract methods.

Additionally, `BasicMenuBar` now has a new action and signal to close
the current window to return to the centralised one.