Histalign/Registration+Backend: Add caching to interpolation
This is a bit of pain since the current workflow doesn't provide
identifying information so caching is not enabled by default.
In order to enable it, the user needs to provide the alignment directory
which allows creating a hash from it. This means the interpolated volume
cache can easily be cross-referenced with the alignment volume cache.
In addition, if the interpolation is limited to a certain mask, the user
can (or rather has to if they want to cache) provide a mask name which
does not change the hash but allows individual cached file for each
masked region.

This might change in the future. However, for now, masking a region does
not return the same average fluorescence as does interpolating the whole
volume. Once this has been solved, the interpolation cache might be a
single file as well.