Histalign/Registration+CommonWidgets: Add shortcut-aware widgets
As the name suggests, this allows widgets to be aware of their own
shortcut and to proudly displays it in their tooltip.

Note that the inheritance is a bit ugly but PySide just doesn't allow
easy multiple inheritance, hence no way to have a single
`ShortcutAwareWidget` that inherits `QWidget` and is then inherited
by the buttons.
Also, overriding `setShortcut` would duplicate the code in each class
that wants to be aware.