Histalign/Registration: Fix thumbnails bunching up on project load
When loading a previously created project with a large number of images,
the thumbnails would be bunched up at the top of the scroll area and
the only way to fix it was to resize the window at least once.
This instead creates a very large scroll area that can easily handle the
thumbnails and is then automatically resized before being shown to the
user.