Histalign/Registration: Add LUT functionality
It is now possible to apply a basic LUT to the image being aligned.
This has had the added benefit of reworking the way in which alpha is
handled, simplifying everything by simply converting the image to an
indexed format instead of simply grayscale. This allows the alpha to
be propagated through the LUT, meaning the original image never needs
to change in memory nor does a copy need to be made every time the alpha
changes.