Histalign/Registration: Add option to clear new project directory
If the directory is not empty, the user receives a warning, asking them
if they truly want to use this directory for the project. If they decide
to go ahead, the contents are cleared and the project directory is used.
Otherwise, the contents are left intact and the user can pick another
directory.