Histalign/Workspace+CCF: Add support for annotation volumes
This provides an API similar to the atlas volume one.
The main difference is the way in which the annotated volume is wrapped.
In order to reduce the mmeory footprint of the volume after loading,
the volume values are remapped to a range that fits inside a uint16
array instead of the original uin32 array. This is gained from the fact
that the Allen Institute has reserved some ranges and therefore has many
gaps in the annotated volume values.