Histalign/Workspace: Account for loss of precision with padding
Since we compute the `slice_plane` origin with integer precision, try to
account for the discrepancy in plane centre which is computed with
float precision.
This results in a slightly smoother transition between different pitch
when there is an offset. Additionally, changing resolution does not
randomly shift the image towards the top left (since we're not padding
when offset, pitch, and yaw are all zero, the previous starting 1-pixel
padding which was a smaller padding for higher resolutions is no longer
relevant).