Histalign/Quantification: Revamp GUI and add cortex depth measurement
The GUI now has a dynamic animated widget that is relevant to the
currently-selected measure. This way, it is easier to only show the
relevant parameters to the user (rather than disable the irrelevant
ones).

Additionally, this adds the skeleton for a cortical depth measurement
widget (to provide something to change between).