Histalign/Registration: Fix `build_alignment`volume`
Since the recent change to the way slices are centered during the
frontend alignment, there was a shape mismatch between the registered
slice and the plane mesh.
This fixes that by adding some info regarding the padding into the mesh
metadata. This information is then used to crop the registered slice
back down to a shape compatible with the mesh.