Histalign/Registration+Workspace: Make volume loading cancellable
Previously, only using `QThread`s to do all the volume-loading work led
to them being hard to stop. They could not periodically check a stop
flag since they were doing all the work. They now use processes to
download and then load the volumes. Because processes are much easier to
terminate (and don't freeze the whole app randomly), they are handy to
allow the user to cancel the loading of a project.