Histalign/Registration: Allow quantification of Z-stacks
This provides the ability ot handle Z-stacks.
Currently, the alignment assumes the stack has the same spacing as the
alignment volume. In the future, this information should be requested
from the user when either opening an image directory or when running
quantification.

The aligned of a Z-stack works by assuming the alignment was done on a
projection that sits in the middle of the stack. From there, each slice
away from that middle is snapped to the closest grid point based on the
stack spacing. Currently, no translation is happening to correct for the
difference between the true centre of the slice and the snapping point.

Once a snapping point has been found for each of the slices, the stack
is sub-projected so that slices being snapped to the same point are
projected as their own mini-stack.

Finally, once all the images have been sub-projected, they are inserted
at the snap point as would happen with a simple 2D image.