Histalign/Registration+Maths: Add landmark registration GUI
The GUI allows the user to set a number of points on the atlas and point
to where they are on the alignment image.
The current requirement is of 9 points at least.

The user is free to zoom and pan around both images, and can preview the
transformation result from the current points.
Once they are satisfied, the user can apply the current transformation
to the main GUI.