Histalign/Misc: Rework structure models and display
This is a big commit that should have been multiple smaller ones.
There are two main changes. The way in which structures are organised
into model and view has been changed to a more straightforward approach.
This provides a nicer, two-column view which allows searching by acronym
or by name by the user.

The quantification (and the QA) GUI has been revamped to be clearer and
to make use of the new structure tree view.