Histalign/Registration: Revamp thumbnails widgets
This provides dynamic resizing of the thumbnails scroll area, meaning
the thumbnails resize to "optimal" sizes, switching the number of
columns to try to maintain an "optimal" size.
There is also more feedback in the way thumbnails are displayed when
active (image currently in aligner), selected (single left-click), and
highlighted (mouse hovering).
Additionally, the file name is now nicely displaying below the thumbnail
instead of in a status bar at the bottom of the scroll area. This makes
ordering thumbnails and figuring out the expected offset much more user-
friendly.
Finally, drag & drop now displays better feedback by attaching a grab
of the source widget during the drag.

Thumbnail adjacent changes are mostly limited to changing the signature
of `HistologySlice.generate_thumbnail` to return the cache location of
the thumbnail rather than passing it around through signals as a NumPy
array. Previously, the cache was optional but it has since become
mandatory, meaning that passing the file name is sufficient to load the
file. This probably means a slight slowdown when initially generating
the thumbnails since we save them then load the back immediately but
should be a bigger speed-up when we get a cache hit (e.g., when loading
a project for the second time).