Histalign/Misc: Rework volume slicing
This new approach provides a way to both centre the image in the view
as well as avoid any cropping that was happening before with extreme
angles.
This provides its own way of cropping and padding the image from the
slicing plane so that the display image is centred on the origin passed
to `vedo.Volume.slice_plane`.