Histalign/Visualisation: Add draft rework of the visualisation GUI
The plan is to have a navigation widget on the left to provide the user
a way to open images or brain volumes.
On the right, the information widget will allow selecting structures
and seeing details about them relevant to the opened file.

The central view will display either a single image (2D slice) or a
3D "FMRI" view of a 3D volume.