Histalign/Registration: Speed up registrator and support large images
OpenCV has open issues dating back at least a decade on the limitations
of its warping functions which can only handle images that have
dimensions small enough to fit in a short (linked to SIMD so not an
easy fix). To avoid falling over when we encounter such an image,
we now fall back to PIL to do the warping. It's not much of a memory
difference (we can share the memory between NumPy and PIL) but the
routine is about 10x slower for large images.

Additionally, avoid unnecessary transpositions when rescaling. This also
provides a slight speed improvement.