Histalign/Registration: Fix contour generation and clear up functions
There is now a way to request the contours of a structure from a
`Registrator`. This contour is generated from a low resolution image
of the volume (i.e., resolution of the slice) and then upscaled and
transformed as appropriate.

Additionally, clarity has been improved with regards to the forward/
reverse and invert documentation and arguments.
Previously, invert needed to be set to `True` when doing a forward
registration because skimage needs the inverse map. However, this
can be confusing when working with the module, hence the invert logic
has been aligned with the forward/reverse registration. A forward
registration (reproducing the registration from the GUI) now requires
invert to be `False` and a reverse registration (registering the volume
to the histology) requires invert to be `True`.