Histalign/Visualisation: Improve navigation widgets
The GUI is now a lot cleaner, providing a (hopefully) more intuitive
interaction with the underlying logic.

Additionally, the navigation logic has been moved to the
`common_widgets` module to allow reusing it elsewhere in the future as
this does not feel visualisation-specific.

Finally, a `DoubleClickMixIn` has been added to allow widgets
to react to double clicks as if they had been clicked normally.
Widgets that already react to clicks (e.g., buttons) simply stop
reacting to single clicks and only react to double clicks. Widgets that
do not react to clicks normally (e.g., labels or custom widgets) now
automatically receive a `clicked` signal that is emitted on double
clicks.