Histalign/Misc: Start migration to a single, tab-based GUI
The previous approach of having a single `QMainWindow` per task was
handy while developing but is a poor experience for end users and can
lead to layout bugs, unstable sizing, confusion, and many opening this,
closing that, opening this again.

The GUI is now a single main window with a single menu bar and status
bar. The four currently available tabs are:
- Registration
- Volume builder
- Quantification
- Visualisation

Currently, only registration has really been integrated into the more
centralised GUI.
The volume builder is also a new addition and will be used to queue
jobs to generate 3D volumes based on alignment. This will mean that
quantification also needs to be adjusted to no longer provide that
functionality.

In addition to centralising everything, this patch provides new icons,
more ergonomic collapsible widgets, and other less important features.