Histalign/Registration: Simply and improve alignment array algorithm
This makes a few changes to the `build_aligned_array` function that
make the code much more readable/maintainable and fix an issue that
was causing inserted slices to be arbitrarily rotated when pitch and
yaw were non-zero.

The first is achieved through a centralised function without running
out into split functions. This makes it much easier to follow what
is actually happening without jumping around the source code.
The second is achieved through using a more straightforward approach
to inserting slices. Instead of doing any kind of sub-projections like
before, each slice of the stack is given its own point cloud from which
its data is interpolated into the final volume.