Histalign/Frontend: Fix pop-up for saving/discarding project
The user was prompted twice when creating or opening a new project from
a dirty one if they chose to discard it at the first pop-up.
Now, the dirty state is temporarily disabled if saved/discarded on the
first pop-up ad the state is restored if the dialog is cancelled. This
way if the user saved, the saved state is conserved and if the user
discarded, they are asked again if they cancelled before finishing
their create/open action.