Histalign/Meta: Embed `histoflow` into `histalign`
This brings over the `histoflow` repo to be used inside of `histalign`.
This is the first step towards having it replace the IO backend.
Further commits will replace interactions with the file system to use
`histoflow` (which is now just IO).

Bringing in `histoflow` also provides its whole CLI to `histalign`,
allowing conversions, etc.