IO: Allow more parameters when opening a file
This commit also marks the passage to the new style. Previously, commit
messages were prefixed with "Histalign" in prevision of `histoflow`
being merged which would have had its own prefix.
However, the decision to merge `histoflow` not as a sub-package but as
an integral part of `histalign` means that they are now the same
package and prefixes are not necessary.