IO: Improve dimension order parsing
Instead of only relying on the guessing heuristic when the dimension
order of a file is unknown, we now try to parse it from metadata before
doing any guess work.

One of the major side effects of this is it is now possible to rearrange
dimensions in files (e.g., rearranging a file to have XY instead of
YX) when converting.