Misc: Allow CCF path retrieval to ensure path is valid
The option to download the files for which paths are requested has been
added to avoid the common pattern of importing the path retrieval
function and the download function, to then have to check whether the
returned path exists and download it if not.
The option to handle to download manually is still possible by not
requesting it be done.

Additionally, the `paths` and `downloads` CCF APIs have been merged
into the __init__ module for simplicity.