Meta: Move `language_helpers` to the root and add `unwrap` helper
This new helper is in preparation for better typing and will help
indicate to mypy that potentially-none values should not be.

In case the GUI logic does have a bug where a `None` ends up where it
shouldn't be at runtime, the `unwrap` helper will catch the error early
enough that the traceback should make it obvious that a `None` was
received when it wasn't expected, without having to have a bunch of:
`if Foo.bar is not None` everywhere.

This is very much inspired from Rust.