pyproject.toml/Typing: Add convenience commands to hunt for errors
The first new command runs the regular check but displays the number of
errors per file rather than showing the errors themselves.
The next one displays the file path with the fewest errors.
The final two display+copy and display (respectively) the output of
`type:check` for the file with fewest errors (and copy the path to said
file).