Registration: Fix signed landmark registration scaling
Previously, it was not possible to extract the sign of the scaling
applied as the result of landmark registration, meaning that images
that needed to be flipped left-right were not once the registration had
been applied (even though it worked fine in the preview before
applying).
Now, the scaling updated on the GUI should be correct when the
landmark registration output requires such flipping.