Docs/Scripts: Add documentation screenshot preprocessing script
When taking screenshots of WSL apps running through an X server (here,
VcXsrv), there is 1 pixel on each size which also captures the
underlying window. In addition, the corners are rounded while the
screenshot captures an image with square corners.
This script can be used in Gimp (Python-fu) to crop the 1-pixel border
around the image and fill in each corner so that the colour difference
is not jarring when viewing the documentation in light mode.