Docs: Remove unnecessary typing in `Returns:` docstring sections
When using mkdocstring/mkdocs-gen-files, the docstrings are expected not
to repeat the returned type if it is already indicated in the function
signature. Since we always include it, all docstrings have been stripped
down to the description of what they return without the extra typing.

Additionally, indentation has been updated so that multi-line
descriptions are considered as one block.