Registration+Workspace: Fix contour generation
Getting the reverse-registered contour previously computed it based off
of a slice that had been cut using linear interpolation and then
potentially rotated with cubic interpolation. This led to a bunch of
different contour artifacts since some smaller contours we detected
around the main one caused by the interpolation.

This commit leaves the option to still do that but makes the default use
nearest interpolation, leading to better contours.